feat: simplify station deletion by removing confirmation modal
