+++ src/components/side-panel/station-side-panel/station-side-panel-footer.tsx
@@ -1,20 +1,39 @@
-import { Button, HStack } from '@chakra-ui/react';
+import { Button, HStack, useToast } from '@chakra-ui/react';
 import { RmgSidePanelFooter } from '@railmapgen/rmg-components';
-import { useState } from 'react';
 import { useRootDispatch, useRootSelector } from '../../../redux';
-import RemoveConfirmModal from '../../modal/remove-confirm-modal';
 import { setCurrentStation, setLoopMidpointStation } from '../../../redux/param/param-slice';
 import { useTranslation } from 'react-i18next';
-import { RmgStyle } from '../../../constants/constants';
+import { Events, RmgStyle, SidePanelMode } from '../../../constants/constants';
+import { checkStationCouldBeRemoved, removeStation } from '../../../redux/param/remove-station-action';
+import { setSelectedStation, setSidePanelMode } from '../../../redux/app/app-slice';
+import { removeInvalidColineOnRemoveStation } from '../../../redux/param/coline-action';
+import rmgRuntime from '@railmapgen/rmg-runtime';
 
 export default function StationSidePanelFooter() {
     const { t } = useTranslation();
     const dispatch = useRootDispatch();
+    const toast = useToast();
 
     const { selectedStation } = useRootSelector(state => state.app);
     const { loop, style } = useRootSelector(state => state.param);
 
-    const [isRemoveModalOpen, setIsRemoveModalOpen] = useState(false);
+    const handleRemove = () => {
+        const result = dispatch(checkStationCouldBeRemoved(selectedStation));
+        if (result) {
+            dispatch(setSidePanelMode(SidePanelMode.CLOSE));
+            dispatch(setSelectedStation('linestart'));
+            dispatch(removeInvalidColineOnRemoveStation(selectedStation));
+            dispatch(removeStation(selectedStation));
+        } else {
+            toast({
+                title: t('Unable to remove this station.'),
+                status: 'error',
+                duration: 3000,
+                isClosable: true,
+            });
+        }
+        rmgRuntime.event(Events.REMOVE_STATION, { success: result });
+    };
 
     return (
         <RmgSidePanelFooter>
@@ -31,12 +50,10 @@ export default function StationSidePanelFooter() {
                         {t('Set as midpoint')}
                     </Button>
                 )}
-                <Button size="sm" variant="outline" onClick={() => setIsRemoveModalOpen(true)}>
+                <Button size="sm" variant="outline" onClick={handleRemove}>
                     {t('StationSidePanel.footer.remove')}
                 </Button>
             </HStack>
-
-            <RemoveConfirmModal isOpen={isRemoveModalOpen} onClose={() => setIsRemoveModalOpen(false)} />
         </RmgSidePanelFooter>
     );
 }
